REF: Move generic methods to aggregation.py (#30856)
